+++ packages/ethers/src/index.tsx
@@ -13,6 +13,7 @@ export {
   NetworkButton,
   AppKit
 } from '@reown/appkit-scaffold-react-native';
+import type { EventName, EventsControllerState } from '@reown/appkit-scaffold-react-native';
 
 export { defaultConfig } from './utils/defaultConfig';
 
@@ -129,7 +130,7 @@ export function useAppKitError() {
   };
 }
 
-export function useAppKitEvents() {
+export function useAppKitEvents(callback?: (newEvent: EventsControllerState) => void) {
   if (!modal) {
     throw new Error('Please call "createAppKit" before using "useAppKitEvents" hook');
   }
@@ -139,12 +140,30 @@ export function useAppKitEvents() {
   useEffect(() => {
     const unsubscribe = modal?.subscribeEvents(newEvent => {
       setEvents({ ...newEvent });
+      callback?.(newEvent);
     });
 
     return () => {
       unsubscribe?.();
     };
-  }, []);
+  }, [callback]);
 
   return event;
+}
+
+export function useAppKitEventSubscription(
+  event: EventName,
+  callback: (newEvent: EventsControllerState) => void
+) {
+  if (!modal) {
+    throw new Error('Please call "createAppKit" before using "useAppKitEventSubscription" hook');
+  }
+
+  useEffect(() => {
+    const unsubscribe = modal?.subscribeEvent(event, callback);
+
+    return () => {
+      unsubscribe?.();
+    };
+  }, [callback, event]);
 }